requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
prettytable==3.10.0
pyDataverse==0.3.1
python-irodsclient==2.0.0
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
urllib3==2.2.0
wcwidth==0.2.13
